+++ edit.php
@@ -8,41 +8,51 @@
 <body>
 
 <?php
+
+  function BlockSQLInjection($str){
+      return str_replace(array("'",'"'), array("&quot;","&quot;"), $str);
+  }
+
   session_start();
   if (isset($_SESSION['auth']) and $_SESSION['auth'] == true){
     $connection = mysqli_connect('127.0.0.1', 'root', '');
     mysqli_select_db($connection, 'blog');
     mysqli_set_charset($connection, 'utf8');
-    if (isset($_POST["id"]) and isset($_POST["title"]) and isset($_POST["articleText"])){
-      mysqli_query($connection, 'UPDATE `articles` SET `header`= "'.$_POST["title"].'" , `article_text` = "'.$_POST["articleText"].'" WHERE `id` ='.$_POST["id"].';');
+     if (isset($_POST['newPost']) and ($_POST['newPost'] = 'true')){
+       mysqli_query($connection, 'INSERT INTO `articles` (`header`, `article_text`) VALUE ("", "");');
+       $query_result_id = mysqli_query($connection, 'SELECT * FROM `articles` WHERE (`header` = "") and (`article_text` = "");');
+       $_POST["id"] = mysqli_fetch_all($query_result_id)[0][0];
+     }
+    if (isset($_POST["id"]) and isset($_POST["title"]) and isset($_POST["articleText"]) and is_numeric($_POST["id"])){
+      mysqli_query($connection, 'UPDATE `articles` SET `header`= "'.BlockSQLInjection($_POST["title"]).'" , `article_text` = "'.BlockSQLInjection($_POST["articleText"]).'" WHERE `id` ='.BlockSQLInjection($_POST["id"]).';');
     }
-    if (isset($_POST["id"]) and isset($_POST["comment_id"])){
+    if (isset($_POST["id"]) and isset($_POST["comment_id"]) and is_numeric($_POST["id"]) and is_numeric($_POST["comment_id"])){
       mysqli_query($connection,'DELETE FROM `comments` WHERE `id` = '.$_POST["comment_id"].';');
     }
-    if (isset($_POST["id"])){
-        $query_result_articles = mysqli_query($connection, 'SELECT * FROM `articles` WHERE `id` ='.$_POST["id"].';');
+    if (isset($_POST["id"]) and is_numeric($_POST["id"])){
+        $query_result_articles = mysqli_query($connection, 'SELECT * FROM `articles` WHERE `id` ='.BlockSQLInjection($_POST["id"]).';');
         $article = mysqli_fetch_all($query_result_articles)[0];
         echo "<h1 style='text-align: center;'>Изменить статью</h1>";
         echo "
           <form action = 'edit.php' method = 'POST'>
               <div class='form-group' style='margin-left: 2%'>
                   <input type='hidden' name='id' value='$article[0]'>
-                  <input type='text' name='title' value='$article[1]' class='form-control'><br>
-                  <textarea name='articleText' class='form-control' style='height: 70vh;'>$article[2]</textarea><br>
+                  <input type='text' name='title' value='".htmlspecialchars($article[1], ENT_QUOTES, 'UTF-8')."' class='form-control'><br>
+                  <textarea name='articleText' class='form-control' style='height: 70vh;'>".htmlspecialchars($article[2], ENT_QUOTES, 'UTF-8')."</textarea><br>
                   <input type='submit' value='Сохранить изменения' class='btn btn-success'>
               </div>
           </form>
           ";
-        $query_result_comments = mysqli_query($connection, 'SELECT * FROM `comments` WHERE `article_id` ='.$_POST["id"].';');
+        $query_result_comments = mysqli_query($connection, 'SELECT * FROM `comments` WHERE `article_id` ='.BlockSQLInjection($_POST["id"]).';');
         $comments = mysqli_fetch_all($query_result_comments);
         echo '<ul>';
         foreach ($comments as $comment) {
           echo "<li style='list-style-type: none;'>
                   <div class='card-header' style='width: 70%;'>
-                    <p>$comment[2], $comment[4]:</p>
+                    <p>".htmlspecialchars($comment[2], ENT_QUOTES, 'UTF-8').", ".htmlspecialchars($comment[4], ENT_QUOTES, 'UTF-8').":</p>
                   </div>
                   <div class='card' style='width: 70%; padding-left: 2%;'>
-                    <p>$comment[3]</p>
+                    <p>".htmlspecialchars($comment[3], ENT_QUOTES, 'UTF-8')."</p>
                   </div><br>
                   <form action='edit.php' method='POST'>
                       <input type='hidden' name='id' value='$article[0]'>
